Final Version

+++ src/br/com/ftt/ec6/seniorLiving/service/impl/AccommodationServiceImpl.java
@@ -8,14 +8,17 @@ import br.com.ftt.ec6.seniorLiving.DAO.AccommodationDAO;
 import br.com.ftt.ec6.seniorLiving.DAO.impl.AccommodationDAOImpl;
 import br.com.ftt.ec6.seniorLiving.db.Database;
 import br.com.ftt.ec6.seniorLiving.entities.Accommodation;
+import br.com.ftt.ec6.seniorLiving.entities.Elderly;
 import br.com.ftt.ec6.seniorLiving.entities.RestHome;
 import br.com.ftt.ec6.seniorLiving.exception.AccommodationException;
 import br.com.ftt.ec6.seniorLiving.service.AccommodationService;
+import br.com.ftt.ec6.seniorLiving.service.ElderlyService;
 
 public class AccommodationServiceImpl implements AccommodationService {
 
 	private static AccommodationServiceImpl instance;
 	private AccommodationDAO accommodationDAO = AccommodationDAOImpl.getInstance();
+	private ElderlyService elderlyService = ElderlyServiceImpl.getInstance();
 	
 	private AccommodationServiceImpl() {}
 	
@@ -30,7 +33,7 @@ public class AccommodationServiceImpl implements AccommodationService {
 	public Accommodation save(String name, String description, RestHome restHome) throws AccommodationException {
 		validateAccommodationParameters(name, description);
 		
-		Accommodation similarAccommodation = getAccommodationByName(name);
+		Accommodation similarAccommodation = getAccommodationByNameAndRestHome(name, restHome);
 		
 		if(similarAccommodation != null) { throw new AccommodationException("Acomodação já existe"); }
 		
@@ -56,11 +59,11 @@ public class AccommodationServiceImpl implements AccommodationService {
 	public Accommodation update(Accommodation accommodationUpdate) throws AccommodationException {
 		if(accommodationUpdate == null || accommodationUpdate.getName().trim().isEmpty()) { throw new AccommodationException("Acomodação inválida"); }
 		
-		if(accommodationUpdate.getDescription().trim().isEmpty()) { throw new AccommodationException("Descrição inválida"); }
+		validateAccommodationParameters(accommodationUpdate.getName(), accommodationUpdate.getDescription());
 		
-		Accommodation similarAccommodation = getAccommodationByName(accommodationUpdate.getName());
+		Accommodation similarAccommodation = getAccommodationByNameAndRestHome(accommodationUpdate.getName(), accommodationUpdate.getRestHome());
 		
-		if(similarAccommodation != null && similarAccommodation.getId().equals(accommodationUpdate.getId()) == false) { throw new AccommodationException("Role já existe"); }
+		if(similarAccommodation != null && similarAccommodation.getId().equals(accommodationUpdate.getId()) == false) { throw new AccommodationException("Acomodação já existe"); }
 		
 		EntityManager entityManager = Database.getConnection();
 		entityManager.getTransaction().begin();
@@ -95,7 +98,17 @@ public class AccommodationServiceImpl implements AccommodationService {
 		entityManager.getTransaction().begin();
 		accommodationDAO.startConnection(entityManager);
 		
-		String message = accommodationDAO.delete(id);
+		Accommodation accommodation = accommodationDAO.getById(id);
+		
+		List<Elderly> elderlyList = elderlyService.getElderlyByAccommodation(accommodation);
+		
+		String message = "";
+		
+		if(elderlyList == null || elderlyList.isEmpty()) {
+			message = accommodationDAO.delete(id);
+		}else {
+			message = "Há pacientes vinculados com essa acomodação";
+		}
 		
 		entityManager.getTransaction().commit();
 		entityManager.close();
@@ -132,8 +145,21 @@ public class AccommodationServiceImpl implements AccommodationService {
 		return accommodationList;
 	}
 	
+	private Accommodation getAccommodationByNameAndRestHome(String name, RestHome restHome) {
+		EntityManager entityManager = Database.getConnection();
+		entityManager.getTransaction().begin();
+		accommodationDAO.startConnection(entityManager);
+		
+		Accommodation accommodation = accommodationDAO.getAccommodationByNameAndRestHome(name, restHome);
+		
+		entityManager.close();
+		accommodationDAO.stopConnection();
+		
+		return accommodation;
+	}
+	
 	private void validateAccommodationParameters(String name, String description) throws AccommodationException {
-		if(name == null || name.trim().isEmpty()) { throw new AccommodationException("Acomodação inválida"); }
+		if(name == null || name.trim().isEmpty()) { throw new AccommodationException("Nome inválido"); }
 		
 		if(description == null || description.trim().isEmpty()) { throw new AccommodationException("Descrição inválida"); }
 	}
